+++ RavenorAdventures/Assets/Script/Components System/Components/HealthDamageSystem/CPN_Data_HealthHandler.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface CPN_Data_HealthHandler
 {
-    public float MaxHealth();
+    public int MaxHealth();
     public int MaxArmor();
 
     public int Defense();
